djangoProject2/pyproject.toml
[tool.poetry]
name = "djangohw"
version = "0.1.0"
description = ""
authors = ["drezz11 <Vishnja90@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
django = "^4.1.5"
psycopg2 = "^2.9.5"
pillow = "^9.4.0"
djangorestframework = "^3.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
